+++ ron-test/Words.hs
@@ -9,13 +9,13 @@ import qualified Data.Text.IO as Text
 import           Data.Text.Metrics (levenshteinNorm)
 import           System.IO (print)
 
-import           RON.Data (evalObjectState, execObjectState, newObjectStateWith,
+import           RON.Data (evalObjectState, execObjectState, newObjectFrameWith,
                            reduceObject)
 import           RON.Data.RGA (RgaString)
 import qualified RON.Data.RGA as RGA
 import           RON.Event (applicationSpecific)
 import           RON.Event.Simulation (runNetworkSim, runReplicaSim)
-import           RON.Types (ObjectState)
+import           RON.Types (ObjectFrame)
 
 main :: IO ()
 main = do
@@ -60,15 +60,15 @@ stems word
         | suffix <- ["ed", "es"], Just s <- [Text.stripSuffix suffix word]
         ]
 
-rgaTrick1 :: Text -> Text -> (ObjectState RgaString, ObjectState RgaString)
+rgaTrick1 :: Text -> Text -> (ObjectFrame RgaString, ObjectFrame RgaString)
 rgaTrick1 begin branch1 =
     either (error . show) id .
     runNetworkSim . runReplicaSim (applicationSpecific 1) . runExceptT $ do
-        begin'   <- newObjectStateWith     $ RGA.newFromText begin
+        begin'   <- newObjectFrameWith     $ RGA.newFromText begin
         branch1' <- execObjectState begin' $ RGA.editText branch1
         pure (begin', branch1')
 
-rgaTrick2 :: (ObjectState RgaString, ObjectState RgaString) -> Text -> Text
+rgaTrick2 :: (ObjectFrame RgaString, ObjectFrame RgaString) -> Text -> Text
 rgaTrick2 (begin', branch1') branch2 =
     either (error . show) id .
     runNetworkSim . runReplicaSim (applicationSpecific 2) . runExceptT $ do